refactor(ui): split chat history support
